requirements.txt
kafka-python==2.3.0
redis==5.0.0
duckdb==1.4.4
